+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.2] - 2019-06-18
+### Added
+- Added the splitted login and user button.
+
 ### Changed
 - Changed the ID of the main fluid container (to prevent Bootstrap problems).
 
@@ -82,8 +87,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic mobile style support.
 
 [Unreleased]: https://bitbucket.org/wikiskripta/medik/commits/branch/master
+[1.0.2]: https://bitbucket.org/wikiskripta/medik/commits/tag/v1.0.2
 [1.0.1]: https://bitbucket.org/wikiskripta/medik/commits/tag/v1.0.1
-[1.0.0]: https://bitbucket.org/wikiskripta/medik/commits/tag/v1.0.0 
+[1.0.0]: https://bitbucket.org/wikiskripta/medik/commits/tag/v1.0.0
 [0.6.5]: https://bitbucket.org/wikiskripta/medik/commits/tag/v0.6.5
 [0.6.4]: https://bitbucket.org/wikiskripta/medik/commits/tag/v0.6.4
 [0.6.3]: https://bitbucket.org/wikiskripta/medik/commits/tag/v0.6.3